Move the old JNI enums to a seperate (optional) module - this means the InChI works fine. If you need the enums you can optionally include the cdk-jniinchi-support module.
