+++ storage/jniinchi-support/src/main/java/net/sf/jniinchi/INCHI_RET.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) 2021  John Mayfield
+ *
+ * Contact: cdk-devel@lists.sourceforge.net
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public License
+ * as published by the Free Software Foundation; either version 2.1
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+
+package net.sf.jniinchi;
+
+/**
+ * This class provides backwards compatibility of JNA-INCHI with JNI-INCHI,
+ * this enum was exposed in the CDK API.
+ * @author John Mayfield
+ */
+public enum INCHI_RET {
+    SKIP,
+    EOF,
+    OKAY,
+    WARNING,
+    ERROR,
+    FATAL,
+    UNKNOWN,
+    BUSY;
+}
